# Improve the performance of Clifford.from_circuit()