--- a/test/python/quantum_info/operators/symplectic/test_clifford.py
+++ b/test/python/quantum_info/operators/symplectic/test_clifford.py
@@ -35,7 +35,7 @@ from qiskit.circuit.library import (
 from qiskit.exceptions import QiskitError
 from qiskit.quantum_info import random_clifford
 from qiskit.quantum_info.operators import Clifford, Operator
-from qiskit.quantum_info.operators.symplectic.clifford_circuits import _append_circuit
+from qiskit.quantum_info.operators.symplectic.clifford_circuits import _append_operation
 from qiskit.quantum_info.synthesis.clifford_decompose import (
     decompose_clifford_ag,
     decompose_clifford_bm,
@@ -187,7 +187,7 @@ class TestCliffordGates(QiskitTestCase):
         for gate_name in ("i", "id", "iden", "x", "y", "z", "h", "s", "sdg", "v", "w"):
             with self.subTest(msg="append gate %s" % gate_name):
                 cliff = Clifford([[1, 0], [0, 1]])
-                cliff = _append_circuit(cliff, gate_name, [0])
+                cliff = _append_operation(cliff, gate_name, [0])
                 value_table = cliff.table._array
                 value_phase = cliff.table._phase
                 value_stabilizer = cliff.stabilizer.to_labels()
@@ -221,8 +221,8 @@ class TestCliffordGates(QiskitTestCase):
             with self.subTest(msg="identity for gate %s" % gate_name):
                 cliff = Clifford([[1, 0], [0, 1]])
                 cliff1 = cliff.copy()
-                cliff = _append_circuit(cliff, gate_name, [0])
-                cliff = _append_circuit(cliff, gate_name, [0])
+                cliff = _append_operation(cliff, gate_name, [0])
+                cliff = _append_operation(cliff, gate_name, [0])
                 self.assertEqual(cliff, cliff1)
 
         gates = ["s", "s", "v"]
@@ -232,8 +232,8 @@ class TestCliffordGates(QiskitTestCase):
             with self.subTest(msg="identity for gate %s" % gate_name):
                 cliff = Clifford([[1, 0], [0, 1]])
                 cliff1 = cliff.copy()
-                cliff = _append_circuit(cliff, gate_name, [0])
-                cliff = _append_circuit(cliff, inv_gate, [0])
+                cliff = _append_operation(cliff, gate_name, [0])
+                cliff = _append_operation(cliff, inv_gate, [0])
                 self.assertEqual(cliff, cliff1)
 
     def test_1_qubit_mult_relations(self):
@@ -255,9 +255,9 @@ class TestCliffordGates(QiskitTestCase):
                 split_rel = rel.split()
                 cliff = Clifford([[1, 0], [0, 1]])
                 cliff1 = cliff.copy()
-                cliff = _append_circuit(cliff, split_rel[0], [0])
-                cliff = _append_circuit(cliff, split_rel[2], [0])
-                cliff1 = _append_circuit(cliff1, split_rel[4], [0])
+                cliff = _append_operation(cliff, split_rel[0], [0])
+                cliff = _append_operation(cliff, split_rel[2], [0])
+                cliff1 = _append_operation(cliff1, split_rel[4], [0])
                 self.assertEqual(cliff, cliff1)
 
     def test_1_qubit_conj_relations(self):
@@ -277,10 +277,10 @@ class TestCliffordGates(QiskitTestCase):
                 split_rel = rel.split()
                 cliff = Clifford([[1, 0], [0, 1]])
                 cliff1 = cliff.copy()
-                cliff = _append_circuit(cliff, split_rel[0], [0])
-                cliff = _append_circuit(cliff, split_rel[2], [0])
-                cliff = _append_circuit(cliff, split_rel[4], [0])
-                cliff1 = _append_circuit(cliff1, split_rel[6], [0])
+                cliff = _append_operation(cliff, split_rel[0], [0])
+                cliff = _append_operation(cliff, split_rel[2], [0])
+                cliff = _append_operation(cliff, split_rel[4], [0])
+                cliff1 = _append_operation(cliff1, split_rel[6], [0])
                 self.assertEqual(cliff, cliff1)
 
     @combine(gate_name=("cx", "cz", "swap"), qubits=([0, 1], [1, 0]))
@@ -339,7 +339,7 @@ class TestCliffordGates(QiskitTestCase):
         }
 
         gate_qubits = gate_name + " " + str(qubits)
-        cliff = _append_circuit(Clifford(np.eye(4)), gate_name, qubits)
+        cliff = _append_operation(Clifford(np.eye(4)), gate_name, qubits)
         target = targets_cliffords[gate_qubits]
         self.assertEqual(target, cliff)
 
@@ -351,8 +351,8 @@ class TestCliffordGates(QiskitTestCase):
                 with self.subTest(msg=f"append gate {gate_name} {qubits}"):
                     cliff = Clifford(np.eye(4))
                     cliff1 = cliff.copy()
-                    cliff = _append_circuit(cliff, gate_name, qubits)
-                    cliff = _append_circuit(cliff, gate_name, qubits)
+                    cliff = _append_operation(cliff, gate_name, qubits)
+                    cliff = _append_operation(cliff, gate_name, qubits)
                     self.assertEqual(cliff, cliff1)
 
     def test_2_qubit_relations(self):
@@ -361,50 +361,50 @@ class TestCliffordGates(QiskitTestCase):
         with self.subTest(msg="relation between cx, h and cz"):
             cliff = Clifford(np.eye(4))
             cliff1 = cliff.copy()
-            cliff = _append_circuit(cliff, "h", [1])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "h", [1])
-            cliff = _append_circuit(cliff, "cz", [0, 1])
+            cliff = _append_operation(cliff, "h", [1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "h", [1])
+            cliff = _append_operation(cliff, "cz", [0, 1])
             self.assertEqual(cliff, cliff1)
 
         with self.subTest(msg="relation between cx and swap"):
             cliff = Clifford(np.eye(4))
             cliff1 = cliff.copy()
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "cx", [1, 0])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "swap", [0, 1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "cx", [1, 0])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "swap", [0, 1])
             self.assertEqual(cliff, cliff1)
 
         with self.subTest(msg="relation between cx and x"):
             cliff = Clifford(np.eye(4))
             cliff1 = cliff.copy()
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "x", [0])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "x", [0])
-            cliff = _append_circuit(cliff, "x", [1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "x", [0])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "x", [0])
+            cliff = _append_operation(cliff, "x", [1])
             self.assertEqual(cliff, cliff1)
 
         with self.subTest(msg="relation between cx and z"):
             cliff = Clifford(np.eye(4))
             cliff1 = cliff.copy()
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "z", [1])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "z", [0])
-            cliff = _append_circuit(cliff, "z", [1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "z", [1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "z", [0])
+            cliff = _append_operation(cliff, "z", [1])
             self.assertEqual(cliff, cliff1)
 
         with self.subTest(msg="relation between cx and s"):
             cliff = Clifford(np.eye(4))
             cliff1 = cliff.copy()
-            cliff = _append_circuit(cliff, "cx", [1, 0])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "s", [1])
-            cliff = _append_circuit(cliff, "cx", [0, 1])
-            cliff = _append_circuit(cliff, "cx", [1, 0])
-            cliff = _append_circuit(cliff, "sdg", [0])
+            cliff = _append_operation(cliff, "cx", [1, 0])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "s", [1])
+            cliff = _append_operation(cliff, "cx", [0, 1])
+            cliff = _append_operation(cliff, "cx", [1, 0])
+            cliff = _append_operation(cliff, "sdg", [0])
             self.assertEqual(cliff, cliff1)
 
     def test_barrier_delay_sim(self):