--- a/test/python/quantum_info/operators/symplectic/test_clifford.py
+++ b/test/python/quantum_info/operators/symplectic/test_clifford.py
@@ -422,6 +422,15 @@ class TestCliffordGates(QiskitTestCase):
         value = Clifford(circ)
         self.assertEqual(value, target)
 
+    def test_from_circuit_with_conditional_gate(self):
+        """Test initialization from circuit with conditional gate."""
+        qc = QuantumCircuit(2, 1)
+        qc.h(0).c_if(0, 0)
+        qc.cx(0, 1)
+
+        with self.assertRaises(QiskitError):
+            Clifford(qc)
+
 
 @ddt
 class TestCliffordSynthesis(QiskitTestCase):