### Summary
Improve the performance of Clifford.from_circuit() by avoiding unnecessary `to_instruction` calls.

### Details and comments
As a by-product, the `clifford_circuits._append_circuit` function is refactored into two functions `clifford_circuits._append_circuit` and `clifford_circuits._append_operation` for readability.