@itoko 2022-08-19
Performance improvement by this PR:

[1q-Clifford circuits test case]
With this PR: 432 ms ± 23 ms per loop (mean ± std. dev. of 20 runs, 1 loop each)
main branch: 1.49 s ± 48.6 ms per loop (mean ± std. dev. of 20 runs, 1 loop each)
[2q-Clifford circuits test case]
With this PR: 267 ms ± 15.5 ms per loop (mean ± std. dev. of 20 runs, 1 loop each)
main branch: 626 ms ± 32.8 ms per loop (mean ± std. dev. of 20 runs, 1 loop each)

Test code:
```
# common
from qiskit.circuit import QuantumCircuit
from qiskit.quantum_info import Clifford
from qiskit.quantum_info.random import random_clifford
import numpy as np
rng = np.random.default_rng(seed=123)
```
```
# 1q-Clifford circuits test case
circuits = [random_clifford(1, rng).to_circuit() for _ in range(10000)]
%timeit -r 20 [Clifford.from_circuit(circ) for circ in circuits]
```
```
# 2q-Clifford circuits test case
circuits = [random_clifford(2, rng).to_circuit() for _ in range(2000)]
%timeit -r 20 [Clifford.from_circuit(circ) for circ in circuits]
```