@alexanderivrii 2022-09-12
Thanks! This looks good to me. My original question to @ShellyGarion (which somehow got me assigned to this issue) is what happens when an `Instruction` without `definition` is appended to a `Clifford`, and now I see that it throws a `QiskitError` -- which totally makes sense (as it should be the responsibility of the caller of `_append_operation` to make sure that the operation can be added). 

I have just one tiny comment. In `_append_circuit` you are checking that no instruction in the circuit can have classical bits:
```
for instruction in circuit:
        if instruction.clbits:
            raise QiskitError(
                f"Cannot apply Instruction with classical bits: {instruction.operation.name}"
            )
```
Would it also make sense to check that no instruction is a conditional gate, as for instance for the following circuit
```
qc = QuantumCircuit(2, 1)
qc.h(0).c_if(0, 0)
qc.cx(0, 1)
```