@itoko 2022-10-04
I think this is now ready to be merged. I rebased to the main branch to merge #7269 and added a check for conditional gate following reviewers suggestions.
Could you take a final look, @alexanderivrii @jakelishman ?